+++ src/workers/generateAsync.worker.js
@@ -2,7 +2,7 @@ import {
 	BufferGeometry,
 	BufferAttribute,
 } from 'three';
-import MeshBVH from '../MeshBVH.js';
+import { MeshBVH } from '../core/MeshBVH.js';
 
 global.onmessage = function ( { data } ) {
